gpu: nvgpu: hal: fix compile error of new compile flags
It's preparing to add bellow CFLAGS:
    -Werror -Wall -Wextra \
    -Wmissing-braces -Wpointer-arith -Wundef \
    -Wconversion -Wsign-conversion \
    -Wformat-security \
    -Wmissing-declarations -Wredundant-decls -Wimplicit-fallthrough

Jira GVSCI-11640

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2555057
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ drivers/gpu/nvgpu/hal/cbc/cbc_gm20b.c
@@ -1,7 +1,7 @@
 /*
  * GM20B CBC
  *
- * Copyright (c) 2019-2021, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2019-2022, NVIDIA CORPORATION.  All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -221,12 +221,12 @@ u32 gm20b_cbc_fix_config(struct gk20a *g, int base)
 	u32 val = gk20a_readl(g, ltc_ltcs_ltss_cbc_num_active_ltcs_r());
 
 	if (val == 2U) {
-		return base * 2;
+		return (u32)(base * 2);
 	} else if (val != 1U) {
 		nvgpu_err(g, "Invalid number of active ltcs: %08x", val);
 	}
 
-	return base;
+	return (u32)base;
 }
 
 
@@ -268,7 +268,7 @@ void gm20b_cbc_init(struct gk20a *g, struct nvgpu_cbc *cbc)
 	/* Bug 1477079 indicates sw adjustment on the posted divided base. */
 	if (g->ops.cbc.fix_config != NULL) {
 		compbit_base_post_divide =
-			g->ops.cbc.fix_config(g, compbit_base_post_divide);
+			g->ops.cbc.fix_config(g, (int)compbit_base_post_divide);
 	}
 
 	gk20a_writel(g, ltc_ltcs_ltss_cbc_base_r(),
